Move NotSSB192 into pipeline_utils and setup v2
