+++ src/SOPRANO/pipeline_utils.py
@@ -8,6 +8,7 @@ from SOPRANO.prepare_coordinates import (
     _exclude_positively_selected_genes_disabled,
     _get_protein_complement,
     _non_randomized,
+    _prep_not_ssb192,
     _prep_ssb192,
     _randomize_with_target_file,
     _sort_excluded_regions_for_randomization,
@@ -294,7 +295,23 @@ class UseSSB192(_SSB192Selection):
 
 
 class UseSSB1922(_SSB192Selection2):
-    msg = "Preparing CDS coordinates using 192 substitution"
+    msg = "Preparing CDS coordinates using SSB 192 substitution"
 
     def _apply(self, params: Parameters):
         _prep_ssb192(params)
+
+
+class NotSSB192(_SSB192Selection):
+    """Does not apply ssb192 selection in CDS coordiante prep"""
+
+    @staticmethod
+    def apply(params: Parameters):
+        _SSB192Selection.check_ready(params)
+        _prep_not_ssb192(params)
+
+
+class NotSSB1922(_SSB192Selection2):
+    msg = "Preparing CDS coordinates using SSB 7 substitution"
+
+    def _apply(self, params: Parameters):
+        _prep_not_ssb192(params)
